refactor(crane-mcp): split mcp-setup.ts by agent concern (578 LOC → 3 modules)
- mcp-setup-claude.ts (360 LOC): Claude-specific setup — project trust,
  deny rules, parallel-isolation hooks, .mcp.json sync
- mcp-setup-agents.ts (178 LOC): Gemini, Codex, Hermes MCP registration
- mcp-setup.ts (67 LOC): checkMcpBinary + checkMcpSetup dispatcher + barrel

All callers import from mcp-setup.ts unchanged — barrel preserves public API.
